# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Blickling Fieldbook API
servers:
  - url: /api
    description: Base API path
...
security:
  - cookieAuth: []
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (administrator only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Categories
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "401":
          description: Unauthorized

  # Observation Images
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: blickling.sid
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, active]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [administrator, manager, team_member]
        active:
          type: boolean
        propertyId:
          type: integer
        createdAt:
          type: string
        lastLoginAt:
          type: ["string", "null"]
    UserInput:
      type: object
      required: [name, email, role, password]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [administrator, manager, team_member]
        password:
          type: string
          minLength: 14
          maxLength: 128
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [administrator, manager, team_member]
        active:
          type: boolean
        password:
          type: string
          minLength: 14
          maxLength: 128
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          type: object
          properties:
            name:
              type: string
            size:
              type: integer
            contentType:
              type: string

    # Images
...